Rebased analytics utils into libs/analytics
Util functions should be side-effect free
